requirements.txt
flask>=2.0
flask-cors>=4.0
openai>=1.0
python-dotenv>=1.0
